+++ CHANGES.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 - Support for deleting Odinson Documents (and all associated lucene documents capturing metadata, etc.) from an index via `OdinsonIndex.deleteOdinsonDoc(odinsonDocId)`.
+- Support for updating Odinson Documents (and all associated lucene documents capturing metadata, etc.) in an index via `OdinsonIndex.updateOdinsonDoc(doc)`.
 - `OdinsonIndex.usingIndex(conf: Config)` context manager (ensure an index gets closed)
 - `ExtractorEngine.usingEngine(conf: Config)` context manager (ensure an engine's index and state get closed)